analysis/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analysis"
version = "0.1.0"
description = "video-ai analysis pipeline"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.9",
  "librosa>=0.10",
  "soundfile>=0.12",
  "webrtcvad>=2.0.10",
  "scenedetect>=0.6.4",
  "tqdm>=4.66",
  "pandas>=2.1",
  "pyarrow>=15.0",
]

[project.optional-dependencies]
faces = ["mediapipe>=0.10"]
objects = ["ultralytics>=8.1"]
embed = ["open_clip_torch>=2.24", "torch>=2.2", "Pillow>=10.0"]
asr = ["openai-whisper>=20231117"]
classifier = ["xgboost>=2.0", "scikit-learn>=1.4"]
store = ["redis>=5.0"]
ml = [
  "mediapipe>=0.10",
  "ultralytics>=8.1",
  "open_clip_torch>=2.24",
  "torch>=2.2",
  "Pillow>=10.0",
  "openai-whisper>=20231117",
  "scikit-learn>=1.4",
]

[project.scripts]
analysis = "analysis.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
